requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0

# Visualization
plotly>=5.14.0
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Financial Data
yfinance>=0.2.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# AI/NLP - BART Summarization (for evidence titles)
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Text Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Web/HTTP
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
